Pursue my anti-robots.txt crusade
Look, I get it, robots.txt was here first, but it's time to admit that Google and Bing disagree with us about what it means. It does *not* prevent a page from appearing in Google or Bing. It *only* prevents Google and Bing from crawling the page, but if those search engines want to index the page, they will know the URL and will put that in their index anyway. Sure, you could say they're greedy for links, or that they're literalists in their interepretation of robots.txt ("You said 'No robots, right?'"), but whatever they are, the fact remains: 

## Robots.txt does not, I repeat again, NOT prevent a page from being in search engines

Indeed, as indicated in these changes, if you want an item to get out of search engines, that page should *not* be in robots.txt, so that Google and Bing can *explicitly* crawl it. If you want to really go wild, you should also:

 - Include the page in sitemaps.xml (encouraging crawlers further)
 - Block your sitemaps.xml from being indexed (preventing the URL from being disclosed that way). I added this change to Django. 

And, when you change a page from being public to being private by way of the robots meta or HTTP tag, you should also:

 - Add it to robots.txt for a day so that it immediately gets removed
 - A day or so later, encourage Google and Bing to re-crawl the page so they pick up the new tags

Will I win this crusade and change the world's understanding of robots.txt? Time will tell. It appears that for every instance of this I fix, another ten developers are born. 

----------

Credentials: I run a public-facing website with about ten million court documents.